refactor(git): share git subprocess helpers
Extract shared Git command helpers into utils and update automation and github callers to use the common retry-protected implementation.

Closes #1408

Implemented-By: Codex
